requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
mtcnn>=0.1.1
streamlit>=1.28.0
reportlab>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
librosa>=0.10.0
soundfile>=0.12.0
albumentations>=1.3.0
seaborn>=0.12.0
facenet-pytorch>=2.5.3
